fix refactoring and test package
