+++ tests/whoami_test.go
@@ -1,4 +1,4 @@
-package commands
+package tests
 
 import (
 	"io/ioutil"
